Make logged line number match the actual line numbers for Log4j
Log4j gets confused with line numbers if one wraps it's loggers. Since
we're using the Log4j as backend for commons-logging, the logged line
numbers were off, and pointed to the line numbers of the log
wrappers. By calling the log methods directly on the Log4j loggers, we
avoid this issue and get correct line numbers in logs again.

This approach means a good deal of duplication. But by restructuring
the wrapping, we can push duplication down to CommonsLog, and can
re-use that in child classes.