Avoid hanging HttpServer when stopping without having starting it
When starting the server in a try block and stopping it in the
corresponding finally block, an exception would cause to stop the
server without having started it. This hung in the joining of the
servers thread pool. By running the join only if the server has been
started, we avoid the hanging join, and make use of try/finally blocks
easier when starting/stopping a HttpServer.