Add basic Http client
This basic Http client abstracts away sending Http requests, and will
eventually allow easy POSTing of JSON data and other of our common use
cases.

We'd like to use okhttp client, as most of it's fluent interface is
nice already. However, it's lacking some flexibility that we need at
the end of the day. So instead of starting with okhttp and having to
switch everything to HttpClient later on, we start with HttpClient
right away.