Avoid NPEs when getting BuildProperties in stripped JARs
If a JAR is missing the needed properties files to access build
properties, the corresponding InputStream is null, which caused NPEs
when trying to load properties from it.
The NPEs are misleading and more harmful than useful, as the program
can continue just fine without the needed properties files. Of course,
some settings get nulled away, but still. So we switch to logging the
failure, instead of throwing NPEs.