Create Guice injector lazily
Each registrering of a module triggered creating an injector. This
makes lazy injecting more painful than it needs to be, as the early
injectors need not have bindings for all required keys and may
therefore fail injection.
We switch to creating the initial injector only if it's really needed
to avoid such problems.