Merge HandleParams and HttpHandlerUtils
While structurally, the separation into a parameter capsule and the
workflow toolbox is the right approach, this separation makes it
cumbersome to used the framework. For example one always needs to pass
the parameter capsule around, which is often in Http response
handling. To ease using the framework, we merge the two. That makes
invocation denser and more to the point.

It also makes the HttpParameter class bigger. But ease of usage is
more important.