Remove injection race condition in ApplicationRunner test
The outcome of the testStaticRun test in ApplicationRunnerTest was
dependent on the previous use of the injector. If the previous use of
the injector used the default modules, the test passed. If the
previous use of the injector cleaned the available modules it failed.

We reset the Injector to the default modules to get rid of the race
condition.