Add indirection when injecting CommonHttpHandler
In 106e4f8897ab52eb66e19abb1ef82f58334aff06 we switched from indirect
to direct injection of CommonHttpHandler as allowed by the switch in
08aebe98b07c98b133e79d6144fa6e634e33cadd to inject modules lazily in
one go.

However, lazy injection in one go causes issues when modules appearing
later in the chain need to access the injector manually to inject
objects from modules that are earlier in the chain. Guice fails
complaining that re-entry is not allowed. This happens for example if
Log4j instantiates RoverLikeRollingPolicy, which needs to manually
inject the clock, as Log4j is not aware of Guice.

So we (in a follow up commit) will switch from lazy injection in one
go to lazy injection module by module. This works around the Guice
issue. But this gets switch gets in the way of direct
CommonHttpHandler injection. So we switch back again to indirect
injection of CommonHttpHandler to prepare for the Guice workaround.