Add load tracker for Caches
When using Caches as expiring Sets, it's sometime useful if adding to
the set actually added a fresh value or just added an already existing
value again. Guava does not allow that in a thread-safe way without
explicit locking.
We add static loaders to determine whether or not a value is new
without resorting to explicit locking.

Change-Id: Ie29d5851358f3c070a3e532fff77524dec70285f