For structured logging escape \ to \< instead of \\
Escaping \ to \\ was closer to the typical C-like string
escaping. However, it requires state when un-escaping. This
state-requirement makes it hard to un-escape using tools like sed.
To ease using basic tools like sed and get simple unescaping, we
escape \ to \<.
That way, we can unconditionally un-escape things like \n to linebreak
wherever it occurs and do not have to check how many backslashes occur
before \n. So we can un-escape stateless.

Change-Id: Ic088601f5f11a16c85b7e88c69d5c4483f33489b